+++ CHANGELOG.md
@@ -243,7 +243,7 @@ This is the last version supporting Django 1.8, 1.9 and 1.10. Django 1.8 is only
 ## 1.5.0 (2017-04-18)
 _Full disclosure: as of april 1st, 2017, I am an employee of [Zostera](http://zostera.nl/), as such I will continue to maintain and improve django-tables2._
  - Made `TableBase.as_values()` an iterator ([#432](https://github.com/bradleyayers/django-tables2/pull/432) by [@pziarsolo](https://github.com/pziarsolo))
- - Added `JSONField` for data in JSON format.
+ - Added `JSONColumn` for data in JSON format.
  - Added `__all__` in `django_tables2/__init__.py` and `django_tables2/columns/__init__.py`
  - Added a setting `DJANGO_TABLES2_TEMPLATE` to allow project-wide overriding of the template used to render tables (fixes [#434](https://github.com/bradleyayers/django-tables2/issues/434)).